agent/requirements.txt
# Baseball Dugout - Player Analysis Agent

# LangChain + MCP
langchain-mcp-adapters
langchain-openai
langgraph

# Environment
python-dotenv

